refactor: move api.py functions into utils.py
Move the api.py functions into utils.py
