+++ .coveragerc
@@ -7,6 +7,7 @@ branch = True
 fail_under = 100
 exclude_lines =
     if __name__ == .__main__.:
+    if TYPE_CHECKING:
     pragma: no cover
 
 [html]
