+++ reports/htmlcov/index.html
@@ -56,23 +56,23 @@
         <tfoot>
             <tr class="total">
                 <td class="name left">Total</td>
-                <td>755</td>
+                <td>767</td>
                 <td>0</td>
-                <td>11</td>
+                <td>13</td>
                 <td>300</td>
                 <td>0</td>
-                <td class="right" data-ratio="1055 1055">100%</td>
+                <td class="right" data-ratio="1067 1067">100%</td>
             </tr>
         </tfoot>
         <tbody>
             <tr class="file">
                 <td class="name left"><a href="main_py.html">main.py</a></td>
-                <td>36</td>
+                <td>37</td>
                 <td>0</td>
                 <td>2</td>
                 <td>10</td>
                 <td>0</td>
-                <td class="right" data-ratio="46 46">100%</td>
+                <td class="right" data-ratio="47 47">100%</td>
             </tr>
             <tr class="file">
                 <td class="name left"><a href="veracode___init___py.html">veracode/__init__.py</a></td>
@@ -85,30 +85,30 @@
             </tr>
             <tr class="file">
                 <td class="name left"><a href="veracode_api_py.html">veracode/api.py</a></td>
-                <td>354</td>
-                <td>0</td>
-                <td>9</td>
                 <td>128</td>
                 <td>0</td>
-                <td class="right" data-ratio="482 482">100%</td>
+                <td>9</td>
+                <td>2</td>
+                <td>0</td>
+                <td class="right" data-ratio="130 130">100%</td>
             </tr>
             <tr class="file">
                 <td class="name left"><a href="veracode_check_compliance_py.html">veracode/check_compliance.py</a></td>
-                <td>54</td>
+                <td>55</td>
                 <td>0</td>
                 <td>0</td>
                 <td>14</td>
                 <td>0</td>
-                <td class="right" data-ratio="68 68">100%</td>
+                <td class="right" data-ratio="69 69">100%</td>
             </tr>
             <tr class="file">
                 <td class="name left"><a href="veracode_config_py.html">veracode/config.py</a></td>
-                <td>199</td>
+                <td>200</td>
                 <td>0</td>
                 <td>0</td>
                 <td>122</td>
                 <td>0</td>
-                <td class="right" data-ratio="321 321">100%</td>
+                <td class="right" data-ratio="322 322">100%</td>
             </tr>
             <tr class="file">
                 <td class="name left"><a href="veracode_constants_py.html">veracode/constants.py</a></td>
@@ -121,12 +121,21 @@
             </tr>
             <tr class="file">
                 <td class="name left"><a href="veracode_submit_artifacts_py.html">veracode/submit_artifacts.py</a></td>
-                <td>88</td>
+                <td>89</td>
                 <td>0</td>
                 <td>0</td>
                 <td>26</td>
                 <td>0</td>
-                <td class="right" data-ratio="114 114">100%</td>
+                <td class="right" data-ratio="115 115">100%</td>
+            </tr>
+            <tr class="file">
+                <td class="name left"><a href="veracode_utils_py.html">veracode/utils.py</a></td>
+                <td>234</td>
+                <td>0</td>
+                <td>2</td>
+                <td>126</td>
+                <td>0</td>
+                <td class="right" data-ratio="360 360">100%</td>
             </tr>
         </tbody>
     </table>
@@ -138,7 +147,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io">coverage.py v5.0.3</a>,
-            created at 2020-01-16 13:02
+            created at 2020-01-18 00:46
         </p>
     </div>
 </div>
