+++ reports/htmlcov/veracode___init___py.html
@@ -66,7 +66,7 @@
     <div class="content">
         <p>
             <a class="nav" href="index.html">&#xab; index</a> &nbsp; &nbsp; <a class="nav" href="https://coverage.readthedocs.io">coverage.py v5.0.3</a>,
-            created at 2020-01-14 23:22
+            created at 2020-01-18 00:46
         </p>
     </div>
 </div>
