+++ reports/htmlcov/veracode_constants_py.html
@@ -127,7 +127,7 @@
     <div class="content">
         <p>
             <a class="nav" href="index.html">&#xab; index</a> &nbsp; &nbsp; <a class="nav" href="https://coverage.readthedocs.io">coverage.py v5.0.3</a>,
-            created at 2020-01-16 12:56
+            created at 2020-01-18 00:46
         </p>
     </div>
 </div>
