+++ tests/test_utils.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # pylint: disable=too-many-public-methods, too-many-lines
 """
-Unit tests for api.py
+Unit tests for utils.py
 """
 
 # built-ins
@@ -18,7 +18,7 @@ from requests.exceptions import HTTPError, Timeout, RequestException, TooManyRed
 # custom
 from tests import constants as test_constants
 from veracode import constants as veracode_constants
-from veracode import api
+from veracode import utils
 from veracode.api import ResultsAPI, UploadAPI
 
 # Setup a logger
@@ -29,21 +29,21 @@ logging.raiseExceptions = True
 LOG = logging.getLogger(__name__)
 
 
-class TestVeracodeApiFunctions(TestCase):
+class TestVeracodeUtils(TestCase):
     """
-    Test api.py's functions
+    Test utils.py
     """
 
     ## validate tests
     # validate decorator on a Results API function
-    @patch("veracode.api.validate_api")
-    @patch("veracode.api.is_valid_attribute")
+    @patch("veracode.utils.validate_api")
+    @patch("veracode.utils.is_valid_attribute")
     def test_validate_decorator(self, mock_is_valid_attribute, mock_validate_api):
         """
         Test the validate decorator
         """
         # Prereqs to test the validate decorator with a valid ResultsAPI
-        @api.validate
+        @utils.validate
         def test_results_function(
             *, results_api: ResultsAPI, variable: int  # pylint: disable=unused-argument
         ) -> None:
@@ -52,7 +52,7 @@ class TestVeracodeApiFunctions(TestCase):
         results_api = ResultsAPI(app_id=test_constants.VALID_RESULTS_API["app_id"])
 
         # Prereqs to test the validate decorator with a valid UploadAPI
-        @api.validate
+        @utils.validate
         def test_upload_function(
             *, upload_api: UploadAPI, variable: int  # pylint: disable=unused-argument
         ) -> None:
@@ -133,12 +133,12 @@ class TestVeracodeApiFunctions(TestCase):
         # argument causes an exception to be raised
         self.assertRaises(
             ElementTree.ParseError,
-            api.parse_xml,
+            utils.parse_xml,
             content=test_constants.XML_API_INVALID_RESPONSE_XML_ERROR["bytes"],
         )
 
         # Succeed when calling the parse_xml function with valid arguments
-        output = api.parse_xml(
+        output = utils.parse_xml(
             content=test_constants.XML_API_VALID_RESPONSE_XML_ERROR["bytes"]
         )
         self.assertEqual(
@@ -157,7 +157,7 @@ class TestVeracodeApiFunctions(TestCase):
         # Succeed when calling the element_contains_error function, given an
         # argument which contains an error
         self.assertTrue(
-            api.element_contains_error(
+            utils.element_contains_error(
                 parsed_xml=test_constants.XML_API_VALID_RESPONSE_XML_ERROR["Element"]
             )
         )
@@ -165,7 +165,7 @@ class TestVeracodeApiFunctions(TestCase):
         # Return False when calling the element_contains_error function, given
         # an argument which doesn't contain an error
         self.assertFalse(
-            api.element_contains_error(
+            utils.element_contains_error(
                 parsed_xml=test_constants.VALID_UPLOAD_API_CREATEBUILD_RESPONSE_XML[
                     "Element"
                 ]
@@ -175,8 +175,8 @@ class TestVeracodeApiFunctions(TestCase):
     ## http_request tests
     # http_request get 200
     @patch("requests.get")
-    @patch("veracode.api.parse_xml")
-    @patch("veracode.api.element_contains_error")
+    @patch("veracode.utils.parse_xml")
+    @patch("veracode.utils.element_contains_error")
     def test_http_request_get_200(
         self, mock_element_contains_error, mock_parse_xml, mock_get
     ):
@@ -202,7 +202,7 @@ class TestVeracodeApiFunctions(TestCase):
             + "/"
             + endpoint
         )
-        response = api.http_request(verb="get", url=url)
+        response = utils.http_request(verb="get", url=url)
 
         self.assertEqual(
             [response.tag, response.attrib],
@@ -218,8 +218,8 @@ class TestVeracodeApiFunctions(TestCase):
 
     # http_request get httperror
     @patch("requests.get")
-    @patch("veracode.api.parse_xml")
-    @patch("veracode.api.element_contains_error")
+    @patch("veracode.utils.parse_xml")
+    @patch("veracode.utils.element_contains_error")
     def test_http_request_get_httperror(
         self, mock_element_contains_error, mock_parse_xml, mock_get
     ):
@@ -250,13 +250,13 @@ class TestVeracodeApiFunctions(TestCase):
             mock_get.return_value.raise_for_status.side_effect = HTTPError()
 
             self.assertRaises(
-                HTTPError, api.http_request, verb="get", url=url,
+                HTTPError, utils.http_request, verb="get", url=url,
             )
 
     # http_request get connectionerror
     @patch("requests.get")
-    @patch("veracode.api.parse_xml")
-    @patch("veracode.api.element_contains_error")
+    @patch("veracode.utils.parse_xml")
+    @patch("veracode.utils.element_contains_error")
     def test_http_request_get_connectionerror(
         self, mock_element_contains_error, mock_parse_xml, mock_get
     ):
@@ -284,13 +284,13 @@ class TestVeracodeApiFunctions(TestCase):
         mock_get.side_effect = ConnectionError()
 
         self.assertRaises(
-            ConnectionError, api.http_request, verb="get", url=url,
+            ConnectionError, utils.http_request, verb="get", url=url,
         )
 
     # http_request get requestexception
     @patch("requests.get")
-    @patch("veracode.api.parse_xml")
-    @patch("veracode.api.element_contains_error")
+    @patch("veracode.utils.parse_xml")
+    @patch("veracode.utils.element_contains_error")
     def test_http_request_get_requestexception(
         self, mock_element_contains_error, mock_parse_xml, mock_get
     ):
@@ -318,13 +318,13 @@ class TestVeracodeApiFunctions(TestCase):
         mock_get.side_effect = RequestException()
 
         self.assertRaises(
-            RequestException, api.http_request, verb="get", url=url,
+            RequestException, utils.http_request, verb="get", url=url,
         )
 
     # http_request get timeout
     @patch("requests.get")
-    @patch("veracode.api.parse_xml")
-    @patch("veracode.api.element_contains_error")
+    @patch("veracode.utils.parse_xml")
+    @patch("veracode.utils.element_contains_error")
     def test_http_request_get_timeout(
         self, mock_element_contains_error, mock_parse_xml, mock_get
     ):
@@ -352,13 +352,13 @@ class TestVeracodeApiFunctions(TestCase):
         mock_get.side_effect = Timeout()
 
         self.assertRaises(
-            Timeout, api.http_request, verb="get", url=url,
+            Timeout, utils.http_request, verb="get", url=url,
         )
 
     # http_request get toomanyredirects
     @patch("requests.get")
-    @patch("veracode.api.parse_xml")
-    @patch("veracode.api.element_contains_error")
+    @patch("veracode.utils.parse_xml")
+    @patch("veracode.utils.element_contains_error")
     def test_http_request_get_toomanyredirects(
         self, mock_element_contains_error, mock_parse_xml, mock_get
     ):
@@ -386,13 +386,13 @@ class TestVeracodeApiFunctions(TestCase):
         mock_get.side_effect = TooManyRedirects()
 
         self.assertRaises(
-            TooManyRedirects, api.http_request, verb="get", url=url,
+            TooManyRedirects, utils.http_request, verb="get", url=url,
         )
 
     # http_request get error body response
     @patch("requests.get")
-    @patch("veracode.api.parse_xml")
-    @patch("veracode.api.element_contains_error")
+    @patch("veracode.utils.parse_xml")
+    @patch("veracode.utils.element_contains_error")
     def test_http_request_get_error_body_response(
         self, mock_element_contains_error, mock_parse_xml, mock_get
     ):
@@ -421,13 +421,13 @@ class TestVeracodeApiFunctions(TestCase):
         )
 
         self.assertRaises(
-            RuntimeError, api.http_request, verb="get", url=url,
+            RuntimeError, utils.http_request, verb="get", url=url,
         )
 
     # http_request post 200
     @patch("requests.post")
-    @patch("veracode.api.parse_xml")
-    @patch("veracode.api.element_contains_error")
+    @patch("veracode.utils.parse_xml")
+    @patch("veracode.utils.element_contains_error")
     def test_http_request_post_200(
         self, mock_element_contains_error, mock_parse_xml, mock_post
     ):
@@ -458,7 +458,7 @@ class TestVeracodeApiFunctions(TestCase):
             + "/"
             + endpoint
         )
-        response = api.http_request(
+        response = utils.http_request(
             verb="post", url=url, data=data, params=params, headers=headers,
         )
 
@@ -476,8 +476,8 @@ class TestVeracodeApiFunctions(TestCase):
 
     # http_request post httperror
     @patch("requests.post")
-    @patch("veracode.api.parse_xml")
-    @patch("veracode.api.element_contains_error")
+    @patch("veracode.utils.parse_xml")
+    @patch("veracode.utils.element_contains_error")
     def test_http_request_post_httperror(
         self, mock_element_contains_error, mock_parse_xml, mock_post
     ):
@@ -514,7 +514,7 @@ class TestVeracodeApiFunctions(TestCase):
 
             self.assertRaises(
                 HTTPError,
-                api.http_request,
+                utils.http_request,
                 verb="post",
                 url=url,
                 data=data,
@@ -524,8 +524,8 @@ class TestVeracodeApiFunctions(TestCase):
 
     # http_request post connectionerror
     @patch("requests.post")
-    @patch("veracode.api.parse_xml")
-    @patch("veracode.api.element_contains_error")
+    @patch("veracode.utils.parse_xml")
+    @patch("veracode.utils.element_contains_error")
     def test_http_request_post_connectionerror(
         self, mock_element_contains_error, mock_parse_xml, mock_post
     ):
@@ -559,7 +559,7 @@ class TestVeracodeApiFunctions(TestCase):
 
         self.assertRaises(
             ConnectionError,
-            api.http_request,
+            utils.http_request,
             verb="post",
             url=url,
             data=data,
@@ -569,8 +569,8 @@ class TestVeracodeApiFunctions(TestCase):
 
     # http_request post requestexception
     @patch("requests.post")
-    @patch("veracode.api.parse_xml")
-    @patch("veracode.api.element_contains_error")
+    @patch("veracode.utils.parse_xml")
+    @patch("veracode.utils.element_contains_error")
     def test_http_request_post_requestexception(
         self, mock_element_contains_error, mock_parse_xml, mock_post
     ):
@@ -604,7 +604,7 @@ class TestVeracodeApiFunctions(TestCase):
 
         self.assertRaises(
             RequestException,
-            api.http_request,
+            utils.http_request,
             verb="post",
             url=url,
             data=data,
@@ -614,8 +614,8 @@ class TestVeracodeApiFunctions(TestCase):
 
     # http_request post timeout
     @patch("requests.post")
-    @patch("veracode.api.parse_xml")
-    @patch("veracode.api.element_contains_error")
+    @patch("veracode.utils.parse_xml")
+    @patch("veracode.utils.element_contains_error")
     def test_http_request_post_timeout(
         self, mock_element_contains_error, mock_parse_xml, mock_post
     ):
@@ -649,7 +649,7 @@ class TestVeracodeApiFunctions(TestCase):
 
         self.assertRaises(
             Timeout,
-            api.http_request,
+            utils.http_request,
             verb="post",
             url=url,
             data=data,
@@ -659,8 +659,8 @@ class TestVeracodeApiFunctions(TestCase):
 
     # http_request post toomanyredirects
     @patch("requests.post")
-    @patch("veracode.api.parse_xml")
-    @patch("veracode.api.element_contains_error")
+    @patch("veracode.utils.parse_xml")
+    @patch("veracode.utils.element_contains_error")
     def test_http_request_post_toomanyredirects(
         self, mock_element_contains_error, mock_parse_xml, mock_post
     ):
@@ -694,7 +694,7 @@ class TestVeracodeApiFunctions(TestCase):
 
         self.assertRaises(
             TooManyRedirects,
-            api.http_request,
+            utils.http_request,
             verb="post",
             url=url,
             data=data,
@@ -704,8 +704,8 @@ class TestVeracodeApiFunctions(TestCase):
 
     # http_request post error body response
     @patch("requests.post")
-    @patch("veracode.api.parse_xml")
-    @patch("veracode.api.element_contains_error")
+    @patch("veracode.utils.parse_xml")
+    @patch("veracode.utils.element_contains_error")
     def test_http_request_post_error_body_response(
         self, mock_element_contains_error, mock_parse_xml, mock_post
     ):
@@ -740,7 +740,7 @@ class TestVeracodeApiFunctions(TestCase):
 
         self.assertRaises(
             RuntimeError,
-            api.http_request,
+            utils.http_request,
             verb="post",
             url=url,
             data=data,
@@ -765,13 +765,13 @@ class TestVeracodeApiFunctions(TestCase):
 
         for verb in ["put", "patch", "delete", "options", "head", "connect", "trace"]:
             self.assertRaises(
-                ValueError, api.http_request, verb=verb, url=url,
+                ValueError, utils.http_request, verb=verb, url=url,
             )
 
     ## is_valid_attribute tests
     # base_url validation
-    @patch("veracode.api.protocol_is_insecure")
-    @patch("veracode.api.is_valid_netloc")
+    @patch("veracode.utils.protocol_is_insecure")
+    @patch("veracode.utils.is_valid_netloc")
     def test_is_valid_attribute_base_url(
         self, mock_is_valid_netloc, mock_protocol_is_insecure
     ):
@@ -783,7 +783,7 @@ class TestVeracodeApiFunctions(TestCase):
         mock_is_valid_netloc.return_value = True
         mock_protocol_is_insecure.return_value = True
         self.assertFalse(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="base_url", value=test_constants.VALID_RESULTS_API["base_url"]
             )
         )
@@ -793,7 +793,7 @@ class TestVeracodeApiFunctions(TestCase):
         mock_is_valid_netloc.return_value = True
         mock_protocol_is_insecure.return_value = False
         self.assertTrue(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="base_url", value=test_constants.VALID_RESULTS_API["base_url"]
             )
         )
@@ -803,7 +803,7 @@ class TestVeracodeApiFunctions(TestCase):
         mock_is_valid_netloc.return_value = True
         mock_protocol_is_insecure.return_value = False
         self.assertFalse(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="base_url",
                 value=test_constants.INVALID_RESULTS_API_MISSING_DOMAIN["base_url"],
             )
@@ -814,7 +814,7 @@ class TestVeracodeApiFunctions(TestCase):
         mock_is_valid_netloc.return_value = True
         mock_protocol_is_insecure.return_value = False
         self.assertTrue(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="base_url", value=test_constants.VALID_RESULTS_API["base_url"],
             )
         )
@@ -824,7 +824,7 @@ class TestVeracodeApiFunctions(TestCase):
         mock_is_valid_netloc.return_value = False
         mock_protocol_is_insecure.return_value = False
         self.assertFalse(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="base_url", value=test_constants.VALID_RESULTS_API["base_url"],
             )
         )
@@ -834,7 +834,7 @@ class TestVeracodeApiFunctions(TestCase):
         mock_is_valid_netloc.return_value = True
         mock_protocol_is_insecure.return_value = False
         self.assertTrue(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="base_url", value=test_constants.VALID_RESULTS_API["base_url"],
             )
         )
@@ -844,7 +844,7 @@ class TestVeracodeApiFunctions(TestCase):
         mock_is_valid_netloc.return_value = True
         mock_protocol_is_insecure.return_value = False
         self.assertTrue(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="base_url",
                 value=test_constants.VALID_RESULTS_API_WITH_PORT_IN_URL["base_url"],
             )
@@ -856,7 +856,7 @@ class TestVeracodeApiFunctions(TestCase):
         mock_protocol_is_insecure.return_value = False
         self.assertRaises(
             ValueError,
-            api.is_valid_attribute,
+            utils.is_valid_attribute,
             key="base_url",
             value=test_constants.INVALID_RESULTS_API_INVALID_PORT["base_url"],
         )
@@ -867,7 +867,7 @@ class TestVeracodeApiFunctions(TestCase):
         mock_protocol_is_insecure.return_value = False
         self.assertRaises(
             ValueError,
-            api.is_valid_attribute,
+            utils.is_valid_attribute,
             key="base_url",
             value="https://example.com::443/testing/",
         )
@@ -877,7 +877,7 @@ class TestVeracodeApiFunctions(TestCase):
         mock_is_valid_netloc.return_value = True
         mock_protocol_is_insecure.return_value = False
         self.assertFalse(
-            api.is_valid_attribute(key="base_url", value="https://example.com/")
+            utils.is_valid_attribute(key="base_url", value="https://example.com/")
         )
 
         # Fail when calling the is_valid_attribute function with a base_url
@@ -885,7 +885,7 @@ class TestVeracodeApiFunctions(TestCase):
         mock_is_valid_netloc.return_value = True
         mock_protocol_is_insecure.return_value = False
         self.assertFalse(
-            api.is_valid_attribute(key="base_url", value="https://example.com/thing")
+            utils.is_valid_attribute(key="base_url", value="https://example.com/thing")
         )
 
     # version validation
@@ -895,19 +895,25 @@ class TestVeracodeApiFunctions(TestCase):
         """
         # Succeed when calling the is_valid_attribute function with a version
         # that maps a string to a string
-        self.assertTrue(api.is_valid_attribute(key="version", value={"test.do": "1.2"}))
+        self.assertTrue(
+            utils.is_valid_attribute(key="version", value={"test.do": "1.2"})
+        )
 
         # Fail when calling the is_valid_attribute function with a version that
         # maps a string to a float
-        self.assertFalse(api.is_valid_attribute(key="version", value={"test.do": 1.1}))
+        self.assertFalse(
+            utils.is_valid_attribute(key="version", value={"test.do": 1.1})
+        )
 
         # Fail when calling the is_valid_attribute function with a version that
         # maps a float to a string
-        self.assertFalse(api.is_valid_attribute(key="version", value={3.141: "2.718"}))
+        self.assertFalse(
+            utils.is_valid_attribute(key="version", value={3.141: "2.718"})
+        )
 
         # Fail when calling the is_valid_attribute function with a version that
         # is a string
-        self.assertFalse(api.is_valid_attribute(key="version", value="failure"))
+        self.assertFalse(utils.is_valid_attribute(key="version", value="failure"))
 
     # endpoint validation
     def test_is_valid_attribute_endpoint(self):
@@ -917,7 +923,7 @@ class TestVeracodeApiFunctions(TestCase):
         # Succeed when calling the is_valid_attribute function with an endpoint
         # that is a valid string
         self.assertTrue(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="endpoint",
                 value="abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789-._~",
             )
@@ -925,15 +931,15 @@ class TestVeracodeApiFunctions(TestCase):
 
         # Fail when calling the is_valid_attribute function with an endpoint
         # that is an empty string
-        self.assertFalse(api.is_valid_attribute(key="endpoint", value=""))
+        self.assertFalse(utils.is_valid_attribute(key="endpoint", value=""))
 
         # Fail when calling the is_valid_attribute function with an endpoint
         # that is an invalid string
-        self.assertFalse(api.is_valid_attribute(key="endpoint", value=";$"))
+        self.assertFalse(utils.is_valid_attribute(key="endpoint", value=";$"))
 
         # Fail when calling the is_valid_attribute function with an endpoint
         # that is an int
-        self.assertFalse(api.is_valid_attribute(key="endpoint", value=7))
+        self.assertFalse(utils.is_valid_attribute(key="endpoint", value=7))
 
     # app_id validation
     def test_is_valid_attribute_app_id(self):
@@ -942,19 +948,19 @@ class TestVeracodeApiFunctions(TestCase):
         """
         # Succeed when calling the is_valid_attribute function with an app_id
         # that is whole number represented as a string
-        self.assertTrue(api.is_valid_attribute(key="app_id", value="54321"))
+        self.assertTrue(utils.is_valid_attribute(key="app_id", value="54321"))
 
         # Fail when calling the is_valid_attribute function with an app_id that
         # is an int
-        self.assertFalse(api.is_valid_attribute(key="app_id", value=54321))
+        self.assertFalse(utils.is_valid_attribute(key="app_id", value=54321))
 
         # Fail when calling the is_valid_attribute function with an app_id that
         # is a string but not a whole number
-        self.assertFalse(api.is_valid_attribute(key="app_id", value="success"))
+        self.assertFalse(utils.is_valid_attribute(key="app_id", value="success"))
 
         # Fail when calling the is_valid_attribute function with an app_id that
         # is a float
-        self.assertFalse(api.is_valid_attribute(key="app_id", value=543.21))
+        self.assertFalse(utils.is_valid_attribute(key="app_id", value=543.21))
 
     # build_dir validation
     def test_is_valid_attribute_build_dir(self):
@@ -964,12 +970,12 @@ class TestVeracodeApiFunctions(TestCase):
         # Succeed when calling the is_valid_attribute function with a build_dir
         # that is a Path object
         self.assertTrue(
-            api.is_valid_attribute(key="build_dir", value=Path("./path.pdb"))
+            utils.is_valid_attribute(key="build_dir", value=Path("./path.pdb"))
         )
 
         # Fail when calling the is_valid_attribute function with a build_dir
         # that is a string
-        self.assertFalse(api.is_valid_attribute(key="build_dir", value="./path.pdb"))
+        self.assertFalse(utils.is_valid_attribute(key="build_dir", value="./path.pdb"))
 
     # build_id validation
     def test_is_valid_attribute_build_id(self):
@@ -979,7 +985,7 @@ class TestVeracodeApiFunctions(TestCase):
         # Succeed when calling the is_valid_attribute function with a build_id
         # that is a valid string
         self.assertTrue(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="build_id",
                 value="abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789-._~",
             )
@@ -987,15 +993,15 @@ class TestVeracodeApiFunctions(TestCase):
 
         # Fail when calling the is_valid_attribute function with a build_id
         # that is an empty string
-        self.assertFalse(api.is_valid_attribute(key="build_id", value=""))
+        self.assertFalse(utils.is_valid_attribute(key="build_id", value=""))
 
         # Succeed when calling the is_valid_attribute function with a build_id
         # that is an invalid string
-        self.assertFalse(api.is_valid_attribute(key="build_id", value=";$"))
+        self.assertFalse(utils.is_valid_attribute(key="build_id", value=";$"))
 
         # Fail when calling the is_valid_attribute function with a build_id
         # that is an int
-        self.assertFalse(api.is_valid_attribute(key="build_id", value=7))
+        self.assertFalse(utils.is_valid_attribute(key="build_id", value=7))
 
     # sandbox validation
     def test_is_valid_attribute_sandbox(self):
@@ -1004,23 +1010,23 @@ class TestVeracodeApiFunctions(TestCase):
         """
         # Succeed when calling the is_valid_attribute function with a sandbox
         # that is whole number represented as a string
-        self.assertTrue(api.is_valid_attribute(key="sandbox", value="54321"))
+        self.assertTrue(utils.is_valid_attribute(key="sandbox", value="54321"))
 
         # Succeed when calling the is_valid_attribute function with a sandbox
         # that is None (the default)
-        self.assertTrue(api.is_valid_attribute(key="sandbox", value=None))
+        self.assertTrue(utils.is_valid_attribute(key="sandbox", value=None))
 
         # Fail when calling the is_valid_attribute function with a sandbox that
         # is an int
-        self.assertFalse(api.is_valid_attribute(key="sandbox", value=54321))
+        self.assertFalse(utils.is_valid_attribute(key="sandbox", value=54321))
 
         # Fail when calling the is_valid_attribute function with a sandbox that
         # is a string but not a whole number
-        self.assertFalse(api.is_valid_attribute(key="sandbox", value="success"))
+        self.assertFalse(utils.is_valid_attribute(key="sandbox", value="success"))
 
         # Fail when calling the is_valid_attribute function with a sandbox that
         # is a float
-        self.assertFalse(api.is_valid_attribute(key="sandbox", value=543.21))
+        self.assertFalse(utils.is_valid_attribute(key="sandbox", value=543.21))
 
     # scan_all_nonfatal_top_level_modules validation
     def test_is_valid_attribute_scan_all_nonfatal_top_level_modules(self):
@@ -1031,7 +1037,7 @@ class TestVeracodeApiFunctions(TestCase):
         # Succeed when calling the is_valid_attribute function with a
         # scan_all_nonfatal_top_level_modules that is a bool
         self.assertTrue(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="scan_all_nonfatal_top_level_modules", value=False
             )
         )
@@ -1039,7 +1045,7 @@ class TestVeracodeApiFunctions(TestCase):
         # Fail when calling the is_valid_attribute function with a
         # scan_all_nonfatal_top_level_modules that is a string
         self.assertFalse(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="scan_all_nonfatal_top_level_modules", value="string"
             )
         )
@@ -1047,7 +1053,7 @@ class TestVeracodeApiFunctions(TestCase):
         # Succeed when calling the is_valid_attribute function with a
         # scan_all_nonfatal_top_level_modules that is an int
         self.assertFalse(
-            api.is_valid_attribute(key="scan_all_nonfatal_top_level_modules", value=1)
+            utils.is_valid_attribute(key="scan_all_nonfatal_top_level_modules", value=1)
         )
 
     # auto_scan validation
@@ -1057,15 +1063,15 @@ class TestVeracodeApiFunctions(TestCase):
         """
         # Succeed when calling the is_valid_attribute function with an auto_scan
         # that is a bool
-        self.assertTrue(api.is_valid_attribute(key="auto_scan", value=False))
+        self.assertTrue(utils.is_valid_attribute(key="auto_scan", value=False))
 
         # Fail when calling the is_valid_attribute function with an auto_scan
         # that is a string
-        self.assertFalse(api.is_valid_attribute(key="auto_scan", value="string"))
+        self.assertFalse(utils.is_valid_attribute(key="auto_scan", value="string"))
 
         # Succeed when calling the is_valid_attribute function with an auto_scan
         # that is an int
-        self.assertFalse(api.is_valid_attribute(key="auto_scan", value=1))
+        self.assertFalse(utils.is_valid_attribute(key="auto_scan", value=1))
 
     # api_key_id validation
     def test_is_valid_attribute_api_key_id(self):
@@ -1075,13 +1081,13 @@ class TestVeracodeApiFunctions(TestCase):
         # Succeed when calling the is_valid_attribute function with an
         # api_key_id that is 32 characters of hex
         self.assertTrue(
-            api.is_valid_attribute(key="api_key_id", value=secrets.token_hex(16))
+            utils.is_valid_attribute(key="api_key_id", value=secrets.token_hex(16))
         )
 
         # Fail when calling the is_valid_attribute function with an api_key_id
         # that is a 32 digit int
         self.assertFalse(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="api_key_id", value=12345678901234567890123456789012
             )
         )
@@ -1089,7 +1095,7 @@ class TestVeracodeApiFunctions(TestCase):
         # Fail when calling the is_valid_attribute function with an api_key_id
         # that is a 32 character string but contains non-hex characters
         self.assertFalse(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="api_key_id", value="a" * 15 + "z" + "b" * 15 + "g"
             )
         )
@@ -1102,13 +1108,13 @@ class TestVeracodeApiFunctions(TestCase):
         # Succeed when calling the is_valid_attribute function with an
         # api_key_secret that is 128 characters of hex
         self.assertTrue(
-            api.is_valid_attribute(key="api_key_secret", value=secrets.token_hex(64))
+            utils.is_valid_attribute(key="api_key_secret", value=secrets.token_hex(64))
         )
 
         # Fail when calling the is_valid_attribute function with an
         # api_key_secret that is 127 characters of hex
         self.assertFalse(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="api_key_secret", value=secrets.token_hex(63) + "0"
             )
         )
@@ -1117,7 +1123,7 @@ class TestVeracodeApiFunctions(TestCase):
         # api_key_secret that is a 128 character string which contains non-hex
         # characters
         self.assertFalse(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="api_key_secret", value=secrets.token_hex(63) + "zz"
             )
         )
@@ -1130,19 +1136,19 @@ class TestVeracodeApiFunctions(TestCase):
         # Succeed when calling the is_valid_attribute function with an
         # ignore_compliance_status that is a bool
         self.assertTrue(
-            api.is_valid_attribute(key="ignore_compliance_status", value=False)
+            utils.is_valid_attribute(key="ignore_compliance_status", value=False)
         )
 
         # Fail when calling the is_valid_attribute function with an
         # ignore_compliance_status that is a string
         self.assertFalse(
-            api.is_valid_attribute(key="ignore_compliance_status", value="string")
+            utils.is_valid_attribute(key="ignore_compliance_status", value="string")
         )
 
         # Succeed when calling the is_valid_attribute function with an
         # ignore_compliance_status that is an int
         self.assertFalse(
-            api.is_valid_attribute(key="ignore_compliance_status", value=1)
+            utils.is_valid_attribute(key="ignore_compliance_status", value=1)
         )
 
     # loglevel validation
@@ -1153,22 +1159,22 @@ class TestVeracodeApiFunctions(TestCase):
         # Succeed when calling the is_valid_attribute function with a loglevel
         # that is an allowed log level, appropriately formatted
         for log_level in veracode_constants.ALLOWED_LOG_LEVELS:
-            self.assertTrue(api.is_valid_attribute(key="loglevel", value=log_level))
+            self.assertTrue(utils.is_valid_attribute(key="loglevel", value=log_level))
 
         # Fail when calling the is_valid_attribute function with a loglevel
         # that is a lowercase version of an allowed log level
         for log_level in veracode_constants.ALLOWED_LOG_LEVELS:
             self.assertFalse(
-                api.is_valid_attribute(key="loglevel", value=log_level.casefold())
+                utils.is_valid_attribute(key="loglevel", value=log_level.casefold())
             )
 
         # Fail when calling the is_valid_attribute function with a loglevel
         # that is an int
-        self.assertFalse(api.is_valid_attribute(key="loglevel", value=20))
+        self.assertFalse(utils.is_valid_attribute(key="loglevel", value=20))
 
         # Fail when calling the is_valid_attribute function with a loglevel
         # that is both an int and a non-existant log level
-        self.assertFalse(api.is_valid_attribute(key="loglevel", value=1020384))
+        self.assertFalse(utils.is_valid_attribute(key="loglevel", value=1020384))
 
     # workflow validation
     def test_is_valid_attribute_workflow(self):
@@ -1178,7 +1184,7 @@ class TestVeracodeApiFunctions(TestCase):
         # Succeed when calling the is_valid_attribute function with a workflow
         # that is a list of allowed workflows
         self.assertTrue(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="workflow", value=list(veracode_constants.SUPPORTED_WORKFLOWS)
             )
         )
@@ -1186,7 +1192,7 @@ class TestVeracodeApiFunctions(TestCase):
         # Fail when calling the is_valid_attribute function with a workflow
         # that is a set of allowed workflows
         self.assertFalse(
-            api.is_valid_attribute(
+            utils.is_valid_attribute(
                 key="workflow", value=veracode_constants.SUPPORTED_WORKFLOWS
             )
         )
@@ -1199,7 +1205,9 @@ class TestVeracodeApiFunctions(TestCase):
             + ["unsupported_workflow"]
             + list(veracode_constants.SUPPORTED_WORKFLOWS)
         )
-        self.assertFalse(api.is_valid_attribute(key="workflow", value=invalid_workflow))
+        self.assertFalse(
+            utils.is_valid_attribute(key="workflow", value=invalid_workflow)
+        )
 
     # verb validation
     def test_is_valid_attribute_verb(self):
@@ -1209,16 +1217,16 @@ class TestVeracodeApiFunctions(TestCase):
         # Succeed when calling the is_valid_attribute function with a verb that
         # is an allowed verb, appropriately formatted
         for verb in veracode_constants.SUPPORTED_VERBS:
-            self.assertTrue(api.is_valid_attribute(key="verb", value=verb))
+            self.assertTrue(utils.is_valid_attribute(key="verb", value=verb))
 
         # Fail when calling the is_valid_attribute function with a verb that
         # is not an allowed verb
-        self.assertFalse(api.is_valid_attribute(key="verb", value=123.2))
+        self.assertFalse(utils.is_valid_attribute(key="verb", value=123.2))
 
         # Fail when calling the is_valid_attribute function with a verb that
         # is not an allowed verb
         for verb in ["put", "patch", "delete", "options", "head", "connect", "trace"]:
-            self.assertFalse(api.is_valid_attribute(key="verb", value=verb))
+            self.assertFalse(utils.is_valid_attribute(key="verb", value=verb))
 
     # catch-all validation
     def test_is_valid_attribute_catch_all(self):
@@ -1228,7 +1236,7 @@ class TestVeracodeApiFunctions(TestCase):
         # Succeed when calling the is_valid_attribute function with a key that
         # is not handled
         self.assertTrue(
-            api.is_valid_attribute(key="ajfoanweofkwmeofmow", value="alksmfo")
+            utils.is_valid_attribute(key="ajfoanweofkwmeofmow", value="alksmfo")
         )
 
     ## configure_environment tests
@@ -1245,7 +1253,7 @@ class TestVeracodeApiFunctions(TestCase):
         values = {}
         with patch.dict("os.environ", values=values, clear=True):
             self.assertIsNone(
-                api.configure_environment(
+                utils.configure_environment(
                     api_key_id=test_constants.VALID_RESULTS_API["api_key_id"],
                     api_key_secret=test_constants.VALID_RESULTS_API["api_key_secret"],
                 )
@@ -1261,7 +1269,7 @@ class TestVeracodeApiFunctions(TestCase):
         }
         with patch.dict("os.environ", values=values, clear=True):
             self.assertIsNone(
-                api.configure_environment(
+                utils.configure_environment(
                     api_key_id=test_constants.VALID_RESULTS_API["api_key_id"],
                     api_key_secret=test_constants.VALID_RESULTS_API["api_key_secret"],
                 )
@@ -1279,7 +1287,7 @@ class TestVeracodeApiFunctions(TestCase):
         }
         with patch.dict("os.environ", values=values, clear=True):
             self.assertIsNone(
-                api.configure_environment(
+                utils.configure_environment(
                     api_key_id=test_constants.VALID_RESULTS_API["api_key_id"],
                     api_key_secret=test_constants.VALID_RESULTS_API["api_key_secret"],
                 )
@@ -1289,7 +1297,7 @@ class TestVeracodeApiFunctions(TestCase):
         # api_key_id and a valid api_key_secret due to the validate decorator
         self.assertRaises(
             ValueError,
-            api.configure_environment,
+            utils.configure_environment,
             api_key_id=invalid_api_key_id,
             api_key_secret=test_constants.VALID_RESULTS_API["api_key_secret"],
         )
@@ -1299,7 +1307,7 @@ class TestVeracodeApiFunctions(TestCase):
         # decorator
         self.assertRaises(
             ValueError,
-            api.configure_environment,
+            utils.configure_environment,
             api_key_id=test_constants.VALID_RESULTS_API["api_key_id"],
             api_key_secret=invalid_api_key_secret,
         )
@@ -1309,7 +1317,7 @@ class TestVeracodeApiFunctions(TestCase):
         # decorator
         self.assertRaises(
             ValueError,
-            api.configure_environment,
+            utils.configure_environment,
             api_key_id=invalid_api_key_id,
             api_key_secret=invalid_api_key_secret,
         )
@@ -1322,12 +1330,12 @@ class TestVeracodeApiFunctions(TestCase):
         # Succeed when calling the validate_api function, given a
         # properly configured ResultsAPI object
         results_api = ResultsAPI(app_id=test_constants.VALID_RESULTS_API["app_id"])
-        self.assertIsNone(api.validate_api(api=results_api))
+        self.assertIsNone(utils.validate_api(api=results_api))
 
         # Succeed when calling the validate_api function, given a
         # properly configured UploadAPI object
         upload_api = UploadAPI(app_id=test_constants.VALID_UPLOAD_API["app_id"])
-        self.assertIsNone(api.validate_api(api=upload_api))
+        self.assertIsNone(utils.validate_api(api=upload_api))
 
         # Fail when attempting to call the validate_api function, given
         # an improperly configured results_api due to an invalid property
@@ -1335,7 +1343,7 @@ class TestVeracodeApiFunctions(TestCase):
             "app_id"
         ]
         self.assertRaises(
-            ValueError, api.validate_api, api=results_api,
+            ValueError, utils.validate_api, api=results_api,
         )
 
         # Fail when attempting to call the validate_api function, given
@@ -1344,7 +1352,7 @@ class TestVeracodeApiFunctions(TestCase):
             "base_url"
         ]
         self.assertRaises(
-            ValueError, api.validate_api, api=upload_api,
+            ValueError, utils.validate_api, api=upload_api,
         )
 
     ## protocol_is_insecure tests
@@ -1353,14 +1361,14 @@ class TestVeracodeApiFunctions(TestCase):
         Test the protocol_is_insecure function
         """
         # protocol_is_insecure must be passed the protocol
-        self.assertRaises(TypeError, api.protocol_is_insecure)
+        self.assertRaises(TypeError, utils.protocol_is_insecure)
 
         # http is insecure
-        output = api.protocol_is_insecure(protocol="http")
+        output = utils.protocol_is_insecure(protocol="http")
         self.assertTrue(output)
 
         # https is secure
-        output = api.protocol_is_insecure(protocol="https")
+        output = utils.protocol_is_insecure(protocol="https")
         self.assertFalse(output)
 
     ## is_null tests
@@ -1369,38 +1377,38 @@ class TestVeracodeApiFunctions(TestCase):
         Test the is_null function
         """
         # is_null must be passed a value
-        self.assertRaises(TypeError, api.is_null)
+        self.assertRaises(TypeError, utils.is_null)
 
         # Python's null equivalent is None
-        output = api.is_null(value=None)
+        output = utils.is_null(value=None)
         self.assertTrue(output)
 
         # An empty string is not null
-        output = api.is_null(value="")
+        output = utils.is_null(value="")
         self.assertFalse(output)
 
         # A tuple is not null
-        output = api.is_null(value=(1, "2"))
+        output = utils.is_null(value=(1, "2"))
         self.assertFalse(output)
 
         # A dict is not null
-        output = api.is_null(value={"a": {"Test": 123}})
+        output = utils.is_null(value={"a": {"Test": 123}})
         self.assertFalse(output)
 
         # An int is not null
-        output = api.is_null(value=12321)
+        output = utils.is_null(value=12321)
         self.assertFalse(output)
 
         # A string is not null
-        output = api.is_null(value="thisisonlyatest")
+        output = utils.is_null(value="thisisonlyatest")
         self.assertFalse(output)
 
         # A list is not null
-        output = api.is_null(value=["a", "b", "c"])
+        output = utils.is_null(value=["a", "b", "c"])
         self.assertFalse(output)
 
         # A set is not null
-        output = api.is_null(value={"1", 2})
+        output = utils.is_null(value={"1", 2})
         self.assertFalse(output)
 
     ## is_valid_netloc tests
@@ -1410,71 +1418,73 @@ class TestVeracodeApiFunctions(TestCase):
         """
         # Succeed when calling the is_valid_netloc function with a legal
         # netloc, containing no subdomains
-        self.assertTrue(api.is_valid_netloc(netloc="example.com"))
+        self.assertTrue(utils.is_valid_netloc(netloc="example.com"))
 
         # Succeed when calling the is_valid_netloc function with a legal
         # netloc, containing numerous subdomains
         self.assertTrue(
-            api.is_valid_netloc(netloc="i.love.to.use.subdomains.example.com")
+            utils.is_valid_netloc(netloc="i.love.to.use.subdomains.example.com")
         )
 
         # Succeed when calling the is_valid_netloc function with a legal netloc
         # due to a 63 character (max valid length) subdomain
-        self.assertTrue(api.is_valid_netloc(netloc="a" * 63 + ".example.com"))
+        self.assertTrue(utils.is_valid_netloc(netloc="a" * 63 + ".example.com"))
 
         # Succeed when calling the is_valid_netloc function with a valid
         # port in the netloc
-        self.assertTrue(api.is_valid_netloc(netloc="example.com:443"))
+        self.assertTrue(utils.is_valid_netloc(netloc="example.com:443"))
 
         # Fail when calling the is_valid_netloc function with a legal netloc,
         # containing no subdomains, but a user/pass specified in the url
-        self.assertFalse(api.is_valid_netloc(netloc="user:pass@example.com"))
+        self.assertFalse(utils.is_valid_netloc(netloc="user:pass@example.com"))
 
         # Fail when attempting to call the is_valid_netloc function with an
         # illegal netloc based on the characters used
-        self.assertFalse(api.is_valid_netloc(netloc="$$"))
+        self.assertFalse(utils.is_valid_netloc(netloc="$$"))
 
         # Fail when attempting to call the is_valid_netloc function with an
         # illegal netloc due to a 64 character (invalid) subdomain
-        self.assertFalse(api.is_valid_netloc(netloc="a" * 64 + ".example.com"))
+        self.assertFalse(utils.is_valid_netloc(netloc="a" * 64 + ".example.com"))
 
         # Fail when attempting to call the is_valid_netloc function with a
         # non-string value
-        self.assertFalse(api.is_valid_netloc(netloc=12321))
+        self.assertFalse(utils.is_valid_netloc(netloc=12321))
 
         # Fail when calling the is_valid_netloc function with an invalid port
         # in the netloc
-        self.assertFalse(api.is_valid_netloc(netloc="example.com:65536"))
+        self.assertFalse(utils.is_valid_netloc(netloc="example.com:65536"))
 
         # Fail when calling the is_valid_netloc function with an IPv4 address
         # as the netloc, as it is not currently supported
-        self.assertFalse(api.is_valid_netloc(netloc="192.0.2.1"))
-        self.assertFalse(api.is_valid_netloc(netloc="192.0.2.1:443"))
-        self.assertFalse(api.is_valid_netloc(netloc="user:pass@192.0.2.1"))
-        self.assertFalse(api.is_valid_netloc(netloc="user:pass@192.0.2.1:443"))
+        self.assertFalse(utils.is_valid_netloc(netloc="192.0.2.1"))
+        self.assertFalse(utils.is_valid_netloc(netloc="192.0.2.1:443"))
+        self.assertFalse(utils.is_valid_netloc(netloc="user:pass@192.0.2.1"))
+        self.assertFalse(utils.is_valid_netloc(netloc="user:pass@192.0.2.1:443"))
 
         # Fail when calling the is_valid_netloc function with a compressed IPv6
         # address as the netloc, as it is not currently supported
-        self.assertFalse(api.is_valid_netloc(netloc="2001:db8::1"))
-        self.assertFalse(api.is_valid_netloc(netloc="2001:db8::1:443"))
-        self.assertFalse(api.is_valid_netloc(netloc="user:pass@2001:db8::1"))
-        self.assertFalse(api.is_valid_netloc(netloc="user:pass@2001:db8::1:443"))
-        self.assertFalse(api.is_valid_netloc(netloc="[2001:db8::1]"))
-        self.assertFalse(api.is_valid_netloc(netloc="[2001:db8::1]:443"))
-        self.assertFalse(api.is_valid_netloc(netloc="user:pass@[2001:db8::1]"))
-        self.assertFalse(api.is_valid_netloc(netloc="user:pass@[2001:db8::1]:443"))
+        self.assertFalse(utils.is_valid_netloc(netloc="2001:db8::1"))
+        self.assertFalse(utils.is_valid_netloc(netloc="2001:db8::1:443"))
+        self.assertFalse(utils.is_valid_netloc(netloc="user:pass@2001:db8::1"))
+        self.assertFalse(utils.is_valid_netloc(netloc="user:pass@2001:db8::1:443"))
+        self.assertFalse(utils.is_valid_netloc(netloc="[2001:db8::1]"))
+        self.assertFalse(utils.is_valid_netloc(netloc="[2001:db8::1]:443"))
+        self.assertFalse(utils.is_valid_netloc(netloc="user:pass@[2001:db8::1]"))
+        self.assertFalse(utils.is_valid_netloc(netloc="user:pass@[2001:db8::1]:443"))
 
         # Fail when calling the is_valid_netloc function with an IPv6 address
         # as the netloc, as it is not currently supported
-        self.assertFalse(api.is_valid_netloc(netloc="2001:db8:0:0:0:0:0:1"))
-        self.assertFalse(api.is_valid_netloc(netloc="2001:db8:0:0:0:0:0:1:443"))
-        self.assertFalse(api.is_valid_netloc(netloc="user:pass@2001:db8:0:0:0:0:0:1"))
+        self.assertFalse(utils.is_valid_netloc(netloc="2001:db8:0:0:0:0:0:1"))
+        self.assertFalse(utils.is_valid_netloc(netloc="2001:db8:0:0:0:0:0:1:443"))
+        self.assertFalse(utils.is_valid_netloc(netloc="user:pass@2001:db8:0:0:0:0:0:1"))
         self.assertFalse(
-            api.is_valid_netloc(netloc="user:pass@2001:db8:0:0:0:0:0:1:443")
+            utils.is_valid_netloc(netloc="user:pass@2001:db8:0:0:0:0:0:1:443")
         )
-        self.assertFalse(api.is_valid_netloc(netloc="[2001:db8:0:0:0:0:0:1]"))
-        self.assertFalse(api.is_valid_netloc(netloc="[2001:db8:0:0:0:0:0:1]:443"))
-        self.assertFalse(api.is_valid_netloc(netloc="user:pass@[2001:db8:0:0:0:0:0:1]"))
+        self.assertFalse(utils.is_valid_netloc(netloc="[2001:db8:0:0:0:0:0:1]"))
+        self.assertFalse(utils.is_valid_netloc(netloc="[2001:db8:0:0:0:0:0:1]:443"))
         self.assertFalse(
-            api.is_valid_netloc(netloc="user:pass@[2001:db8:0:0:0:0:0:1]:443")
+            utils.is_valid_netloc(netloc="user:pass@[2001:db8:0:0:0:0:0:1]")
+        )
+        self.assertFalse(
+            utils.is_valid_netloc(netloc="user:pass@[2001:db8:0:0:0:0:0:1]:443")
         )
